Fixed the type of speed
The speed parameter was set to be an int, should have been a float

+++ RobotMain/LineFollowToSwitch.cpp
@@ -1,12 +1,11 @@
 #include "LineFollowToSwitch.h"
 
 /* Constructor */
-LineFollowToSwitch::LineFollowToSwitch(int speed) : Command("LineFollowToSwitch") {
+LineFollowToSwitch::LineFollowToSwitch(float speed) : Command("LineFollowToSwitch") {
 	_speed = speed;
 }
 
 void LineFollowToSwitch::initialize() {
-
 }
 
 /**
