+++ RobotMain/LineFollowToSwitch.h
@@ -9,13 +9,13 @@
 
 class LineFollowToSwitch : public Command {
 public:
-	LineFollowToSwitch(int speed);
+	LineFollowToSwitch(float speed);
 	void initialize();
 	void execute();
 	void end();
 	bool isFinished();
 private:
-	unsigned long _speed;
+	float _speed;
 	Drivetrain drivetrain;
 	LineTracker lineTracker;
 	AlignmentDetector alignmentDetector;
